requirements.txt
connexion>=1.0.31
gevent>=1.1b1
PyMySQL==0.7.11
PyYAML==3.12
Keras==2.0.6
Theano==0.9.0
scipy==0.19.1
gensim==2.3.0
